refactor: 불필요한 Hook 삭제
